Split the building of an image from the transferring of an image into separate functions. BuildImage API unchanged, internally calls both. Idea is to resolve issues caused by using same remote to build image as destination remote for image push.
